move Field class to field.py
TODO: if willing - subclass boilerplate code in Field
